♻️ refactor(Main): Extract and centralize step rendering logic with `useStepBody` hook
- Removed inline step rendering from `Main` component and replaced it with `useStepBody` hook for improved modularity and maintainability.
- Simplified `Main` component structure by delegating step-specific logic to the new hook.
