+++ app/exam/[attemptId]/Main.tsx
@@ -1,123 +1,15 @@
 'use client';
-import WelcomeStep from '@/exam/[attemptId]/steps/Welcome/WelcomeStep';
-import ReadingIntroStep from '@/exam/[attemptId]/steps/Reading/ReadingIntroStep';
-import ReadingQuestionStep from '@/exam/[attemptId]/steps/Reading/ReadingQuestionStep';
-import ReadingCompleteStep from '@/exam/[attemptId]/steps/Reading/ReadingCompleteStep';
-import ListeningIntroStep from '@/exam/[attemptId]/steps/Listening/ListeningIntroStep';
-import ListeningStep from '@/exam/[attemptId]/steps/Listening/ListeningStep';
-import ListeningCompleteStep from '@/exam/[attemptId]/steps/Listening/ListeningCompleteStep';
-import WritingIntroStep from '@/exam/[attemptId]/steps/Writing/WritingIntroStep';
-import WritingStep from '@/exam/[attemptId]/steps/Writing/WritingStep';
-import WritingCompleteStep from '@/exam/[attemptId]/steps/Writing/WritingCompleteStep';
-import SpeakingIntroStep from '@/exam/[attemptId]/steps/Speaking/SpeakingIntroStep';
-import SpeakingInstructionsStep from '@/exam/[attemptId]/steps/Speaking/SpeakingInstructionsStep';
-import FinalRecapStep from '@/exam/[attemptId]/steps/Final/FinalRecapStep';
 import { JSX, useEffect } from 'react';
 import { useTimerStore } from '@/state/timerStore';
 import PreventBackNavigation from '@/components/PreventBackNavigation';
-import PermissionStep from '@/exam/[attemptId]/steps/Permission/PermissionStep';
-import SpeakingCompleteStep from '@/exam/[attemptId]/steps/Speaking/SpeakingCompleteStep';
 import { useStepStore } from '@/state/stepStore';
-import ReadingQuestionListStep from '@/exam/[attemptId]/steps/Reading/ReadingQuestionListStep';
-
-function StepBody({ current, next }: { current: number; next: () => void }) {
-  const steps = useStepStore((s) => s.steps);
-  const step = steps[current];
-
-  if (!step) {
-    return <div>Loading step...</div>;
-  }
-
-  switch (step.kind) {
-    case 'welcome':
-      return <WelcomeStep onNextAction={next} />;
-    case 'permission':
-      return <PermissionStep onNextAction={next} />;
-    case 'reading-start':
-      return (
-        <ReadingIntroStep
-          title={step.title}
-          subtitle={step.subTitle}
-          durationMs={step.durationMs}
-          kind={step.kind}
-          onNextAction={next}
-        />
-      );
-    case 'reading-question':
-      return (
-        <ReadingQuestionStep sentence={step.sentence} options={step.options} onNextAction={next} />
-      );
-    case 'reading-question-list':
-      return (
-        <ReadingQuestionListStep
-          questions={step.questions}
-          passage={step.passage}
-          onNextAction={next}
-        />
-      );
-    case 'reading-complete':
-      return <ReadingCompleteStep title={step.title} onNextAction={next} />;
-    case 'listening-start':
-      return (
-        <ListeningIntroStep
-          title={step.title}
-          subtitle={step.subTitle}
-          durationMs={step.durationMs}
-          kind={step.kind}
-          onNextAction={next}
-        />
-      );
-    case 'listening-question':
-      return (
-        <ListeningStep audioUrl={step.audioUrl} questions={step.questions} onNextAction={next} />
-      );
-    case 'listening-complete':
-      return <ListeningCompleteStep title={step.title} onNextAction={next} />;
-    case 'writing-start':
-      return (
-        <WritingIntroStep
-          title={step.title}
-          subtitle={step.subTitle}
-          kind={step.kind}
-          durationMs={step.durationMs}
-          onNextAction={next}
-        />
-      );
-    case 'writing-question':
-      return <WritingStep title={step.title} onNextAction={next} />;
-    case 'writing-complete':
-      return <WritingCompleteStep title={step.title} onNextAction={next} />;
-    case 'speaking-start':
-      return (
-        <SpeakingIntroStep
-          title={step.title}
-          subtitle={step.subTitle}
-          durationMs={step.durationMs}
-          kind={step.kind}
-          onNextAction={next}
-        />
-      );
-    case 'speaking-question':
-      return (
-        <SpeakingInstructionsStep
-          recDurationMs={10000}
-          audioFileUrl={step.audioUrl}
-          onNextAction={next}
-        />
-      );
-    case 'speaking-complete':
-      return <SpeakingCompleteStep onNextAction={next} />;
-    case 'final':
-      return <FinalRecapStep />;
-    default:
-      return <div>Invalid step</div>;
-  }
-}
+import { useStepBody } from '@/hooks/useStepBody';
 
 export default function Main(): JSX.Element {
   const currentStepIndex = useTimerStore((s) => s.currentStepIndex);
   const nextStep = useTimerStore((s) => s.nextStep);
   const { steps, isLoading, error, fetchSteps } = useStepStore();
+  const { StepComponent } = useStepBody({ current: currentStepIndex, next: nextStep });
 
   useEffect(() => {
     if (steps.length === 0) {
@@ -137,7 +29,7 @@ export default function Main(): JSX.Element {
     <>
       <PreventBackNavigation />
       <section className="p-6 text-center">
-        <StepBody current={currentStepIndex} next={nextStep} />
+        <StepComponent />
       </section>
     </>
   );
